+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+- Scanning from outside [scanBarcode]
+
+## 1.0.1
+
+- Add flags [camera, barcode]
+
 ## 1.0.0
 
 - Capture images using the device's camera.